Revert "Задание 6"
