+++ src/TypesOfTasks/managers/InMemoryHistoryManager.java
@@ -5,79 +5,30 @@ import TypesOfTasks.tasks.Task;
 import java.util.*;
 
 public class InMemoryHistoryManager implements HistoryManager {
-    private static class Node {
-        Task task;
-        Node prev;
-        Node next;
-
-        public Node(Node prev, Task task, Node next) {
-            this.prev = prev;
-            this.task = task;
-            this.next = next;
-        }
-    }
-
-    private Node head;
-    private Node tail;
-    private final Map<Integer, Node> nodeMap = new HashMap<>();
-
-    private void linkLast(Task task) {
-        final Node oldTail = tail;
-        final Node newNode = new Node(oldTail, task, null);
-        tail = newNode;
-        if (oldTail == null) {
-            head = newNode;
-        } else {
-            oldTail.next = newNode;
-        }
-        nodeMap.put(task.getId(), newNode);
-    }
-
-    private List<Task> getTasks() {
-        List<Task> tasks = new ArrayList<>();
-        Node current = head;
-        while (current != null) {
-            tasks.add(current.task);
-            current = current.next;
-        }
-        return tasks;
-    }
-
-    private void removeNode(Node node) {
-        if (node == null) return;
-        if (node.prev != null) {
-            node.prev.next = node.next;
-        } else {
-            head = node.next;
-        }
-        if (node.next != null) {
-            node.next.prev = node.prev;
-        } else {
-            tail = node.prev;
-        }
-    }
+    private final Map<Integer, Task> historyMap = new LinkedHashMap<>();
 
     @Override
     public void add(Task task) {
         if (task == null) return;
-        if (nodeMap.containsKey(task.getId())) {
-            removeNode(nodeMap.get(task.getId()));
-            nodeMap.remove(task.getId());
+
+        historyMap.remove(task.getId());
+
+        if (historyMap.size() >= 10) {
+            Integer firstKey = historyMap.keySet().iterator().next();
+            historyMap.remove(firstKey);
         }
-        linkLast(task);
+
+        historyMap.put(task.getId(), task);
+    }
+
+
+    @Override
+    public List<Task> getHistory() {
+        return new ArrayList<>(historyMap.values());
     }
 
     @Override
     public void remove(int id) {
-        Node node = nodeMap.remove(id);
-        if (node != null) {
-            removeNode(node);
-        }
-    }
-
-    @Override
-    public List<Task> getHistory() {
-        return getTasks();
+        historyMap.remove(id);
     }
 }
-
